+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-## v0.0.1 - 2023-05-22
+## v0.0.3 - 2023-05-22
+
+- Added Check to make sure a directory is not added as function source for multiple time.
+- Added function to download repo from git and add its events as source.
+
+## v0.0.2 - 2023-05-22
 
 - Updated Readme.